Quest Replacement
If you think about it, EE is a tech mod. The EMC stands for Energy Mass Csomething that starts with C, and therefore represents the scientific fact that matter is simply solid energy. This energy can be harvested by machines in that mod (though the unexplored scientific properties of glowstone and redstone, mixed with diamonds) and stored as pure energy. When a atomic structure is chosen to replicate, the energy is transformed back into mass through processes again unknown but resulting from the obsidian and diamond composite that composes the composition of the condenser.  When the condenser replicates the example item given to it and scanned to an atomic level, it uses the energy it stores. More energy can be generated from the sun (solar panels) and passed through relays (wires) to the condenser, which it can use to make more energy. In conclusion, I'd like to ask you all to consider this: How can a Alchemical Condensenrter Help HUmans in the Real world And what ther Possible IMpacts of this rEvolotionary Technoelosegy. THank you.